Move navigation structure to seperate file
