+++ src/frontend/src/pages/ProjectPage/pages/ContractPage/index.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { RouteComponentProps, Route } from 'react-router-dom';
 import ContractContext from '../../../../contractContex';
-import { Contract, useHistory, combineUrls } from '@equinor/fusion';
-import { NavigationDrawer, NavigationStructure } from '@equinor/fusion-components';
+import { Contract } from '@equinor/fusion';
+import { NavigationDrawer } from '@equinor/fusion-components';
 import ScopedSwitch from '../../../../components/ScopedSwitch';
 import ContractDetailsPage from './pages/ContractDetailsPage';
 import ManagePersonellPage from './pages/ManagePersonnelPage';
-import { History } from "history";
+import useContractPageNavigationStructure from "./useContractPageNavigationStructure";
 
 type ContractPageMatch = {
     contractId: string;
@@ -16,8 +16,8 @@ type ContractPageProps = RouteComponentProps<ContractPageMatch>;
 
 const useContractFromId = (id: string) => {
     const [contract, setContract] = React.useState<Contract | null>(null);
-    const [isFetchingContract, setIsFetchingContract] = React.useState(false);
-    const [fetchContractError, setFetchContractError] = React.useState<Error | null>(null);
+    const [] = React.useState(false);
+    const [] = React.useState<Error | null>(null);
 
     React.useEffect(() => {
         // TODO: fetch and set contract
@@ -27,51 +27,10 @@ const useContractFromId = (id: string) => {
     return { contract };
 };
 
-const createContractPath = (history: History, contractId: string, path: string) => {
-    const base = history.location.pathname.split("/" + contractId)[0];
-    return combineUrls(base, contractId, path);
-};
-
-const createNavItem = (history: History, contractId: string, title: string, path: string): NavigationStructure => ({
-    id: title,
-    title,
-    type: 'section',
-    isActive: history.location.pathname === createContractPath(history, contractId, path),
-    onClick: () => history.push(createContractPath(history, contractId, path)),
-});
-
-const getNavigationStructure = (history: History, contractId: string): NavigationStructure[] => {
-    return [
-        createNavItem(history, contractId, "General", ""),
-        createNavItem(history, contractId, "Manage personnel", "personnel"),
-        {
-            id: 'manage-mpp',
-            title: 'Manage MPP',
-            type: 'grouping',
-            isOpen: true,
-            navigationChildren: [
-                createNavItem(history, contractId, "Actual MPP", "actual-mpp"),
-                createNavItem(history, contractId, "Active requests", "active-requests"),
-                createNavItem(history, contractId, "Log", "Log"),
-            ]
-        },
-    ];
-}
-
-const useNavigationStructure = (contractId: string) => {
-    const history = useHistory();
-    const [structure, setStructure] = React.useState<NavigationStructure[]>(getNavigationStructure(history, contractId));
-
-    React.useEffect(() => {
-        setStructure(getNavigationStructure(history, contractId));
-    }, [contractId, history.location.pathname]);
-
-    return { structure, setStructure };
-};
 
 const ContractPage: React.FC<ContractPageProps> = ({ match }) => {
     const { contract } = useContractFromId(match.params.contractId);
-    const { structure, setStructure } = useNavigationStructure(match.params.contractId);
+    const { structure, setStructure } = useContractPageNavigationStructure(match.params.contractId);
 
     const contractContext = React.useMemo(() => {
         return contract
